requirements.txt
gradio==2.8.5
matplotlib==3.7.1
numba==0.56.4
numpy==1.22.4
pandas==1.5.3
Pillow==9.5.0
scipy==1.10.1
torch==1.7.0
